moving experiment ui to its own package, and removing some files which are oprime specific
