+++ src/com/github/opensourcefieldlinguistics/fielddb/experimentation/ui/ProductionExperimentActivity.java
@@ -1,10 +1,11 @@
-package com.github.opensourcefieldlinguistics.fielddb.lessons.ui;
+package com.github.opensourcefieldlinguistics.fielddb.experimentation.ui;
 
 
 import com.github.opensourcefieldlinguistics.datacollection.AudioRecorder;
 import com.github.opensourcefieldlinguistics.fielddb.Config;
 import com.github.opensourcefieldlinguistics.fielddb.database.DatumContentProvider;
 import com.github.opensourcefieldlinguistics.fielddb.database.DatumContentProvider.DatumTable;
+import com.github.opensourcefieldlinguistics.fielddb.lessons.ui.DatumFragmentPagerAdapter;
 import com.github.opensourcefieldlinguistics.fielddb.R;
 
 import android.content.Intent;
