+++ src/com/github/opensourcefieldlinguistics/fielddb/experimentation/ui/StoryBookSubExperiment.java
@@ -14,7 +14,7 @@
    limitations under the License.
  */
 
-package com.github.opensourcefieldlinguistics.fielddb.lessons.ui;
+package com.github.opensourcefieldlinguistics.fielddb.experimentation.ui;
 
 import java.io.IOException;
 import java.util.ArrayList;
