+++ src/com/github/opensourcefieldlinguistics/fielddb/experimentation/ui/SubExperimentFragment.java
@@ -1,4 +1,4 @@
-package com.github.opensourcefieldlinguistics.fielddb.lessons.ui;
+package com.github.opensourcefieldlinguistics.fielddb.experimentation.ui;
 
 import android.app.Fragment;
