--- src/com/github/opensourcefieldlinguistics/fielddb/lessons/ui/ExperimentActivity.java
@@ -1,68 +1,0 @@
-package com.github.opensourcefieldlinguistics.fielddb.lessons.ui;
-
-import android.annotation.SuppressLint;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.NavUtils;
-import android.view.MenuItem;
-import com.github.opensourcefieldlinguistics.fielddb.R;
-
-/**
- * An activity representing a single Experiment detail screen. This activity is
- * only used on handset devices. On tablet-size devices, item details are
- * presented side-by-side with a list of items in a
- * {@link ExperimentListActivity}.
- * <p>
- * This activity is mostly just a 'shell' activity containing nothing more than
- * a {@link ExperimentFragment}.
- */
-public class ExperimentActivity extends FragmentActivity {
-
-  @SuppressLint("NewApi")
-@Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    this.setContentView(R.layout.activity_experiment_detail);
-
-    // Show the Up button in the action bar.
-    this.getActionBar().setDisplayHomeAsUpEnabled(true);
-
-    // savedInstanceState is non-null when there is fragment state
-    // saved from previous configurations of this activity
-    // (e.g. when rotating the screen from portrait to landscape).
-    // In this case, the fragment will automatically be re-added
-    // to its container so we don't need to manually add it.
-    // For more information, see the Fragments API guide at:
-    //
-    // http://developer.android.com/guide/components/fragments.html
-    //
-    if (savedInstanceState == null) {
-      // Create the detail fragment and add it to the activity
-      // using a fragment transaction.
-      Bundle arguments = new Bundle();
-      arguments.putString(ExperimentFragment.ARG_ITEM_ID,
-          this.getIntent().getStringExtra(ExperimentFragment.ARG_ITEM_ID));
-      ExperimentFragment fragment = new ExperimentFragment();
-      fragment.setArguments(arguments);
-      this.getSupportFragmentManager().beginTransaction().add(R.id.experiment_detail_container, fragment).commit();
-    }
-  }
-
-  @Override
-  public boolean onOptionsItemSelected(MenuItem item) {
-    switch (item.getItemId()) {
-    case android.R.id.home:
-      // This ID represents the Home or Up button. In the case of this
-      // activity, the Up button is shown. Use NavUtils to allow users
-      // to navigate up one level in the application structure. For
-      // more details, see the Navigation pattern on Android Design:
-      //
-      // http://developer.android.com/design/patterns/navigation.html#up-vs-back
-      //
-      NavUtils.navigateUpTo(this, new Intent(this, ExperimentListActivity.class));
-      return true;
-    }
-    return super.onOptionsItemSelected(item);
-  }
-}
