+++ src/com/github/opensourcefieldlinguistics/javascript/FieldDBWebViewClient.java
@@ -10,7 +10,7 @@ import android.webkit.SslErrorHandler;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
-public class OPrimeWebViewClient extends WebViewClient {
+public class FieldDBWebViewClient extends WebViewClient {
   public void onReceivedSslError(WebView view, SslErrorHandler handler, SslError error) {
     handler.proceed();
   }
